Try to fix toolchain bug [#2636] and kernel ltp bug [#1606]
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@4799 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ uClibc/libc/sysdeps/linux/common/bits/ipc.h
@@ -51,6 +51,6 @@ struct ipc_perm
     __kernel_uid_t	cuid;
     __kernel_gid_t	cgid;
     __kernel_mode_t	mode; 
-    unsigned long	seq;
+    unsigned short	seq;
 };
